+++ examples/min_max_size.rs
@@ -3,11 +3,12 @@ extern crate winit;
 fn main() {
     let mut events_loop = winit::EventsLoop::new();
 
-    let _window = winit::WindowBuilder::new()
-        .with_min_dimensions(400, 200)
-        .with_max_dimensions(800, 400)
+    let window = winit::WindowBuilder::new()
         .build(&events_loop)
         .unwrap();
+
+    window.set_min_dimensions(Some((400, 200)));
+    window.set_max_dimensions(Some((800, 400)));
 
     events_loop.run_forever(|event| {
         println!("{:?}", event);