+++ src/platform/android/mod.rs
@@ -248,6 +248,12 @@ impl Window {
     }
 
     #[inline]
+    pub fn set_min_dimensions(&self, _dimensions: Option<(u32, u32)>) { }
+
+    #[inline]
+    pub fn set_max_dimensions(&self, _dimensions: Option<(u32, u32)>) { }
+
+    #[inline]
     pub fn get_inner_size(&self) -> Option<(u32, u32)> {
         if self.native_window.is_null() {
             None